refactor: 코드리뷰반영 폴더구조수정
