+++ src/entities/calendar/__stories__/private-schedule-badge.stories.tsx
@@ -1,4 +1,4 @@
-import { PrivateScheduleBadge } from '..'
+import { PrivateScheduleBadge } from '../ui'
 
 import type { Meta, StoryObj } from '@storybook/react-vite'
