+++ src/entities/calendar/__stories__/private-schedule-modal.stories.tsx
@@ -1,6 +1,6 @@
 import { useState } from 'react'
 
-import { PrivateScheduleModal } from '..'
+import { PrivateScheduleModal } from '../ui'
 
 import type { Meta, StoryObj } from '@storybook/react-vite'
