+++ src/entities/calendar/ui/appointment-badge.tsx
