+++ src/entities/calendar/ui/index.ts
@@ -1,0 +1,7 @@
+export { default as CalendarCell } from './calendar-cell'
+export { default as HeaderMonthLabel } from './header-month-label'
+export { default as HeaderButton } from './header-button'
+export { default as PrivateScheduleBadge } from './private-schedule-badge'
+export { default as PrivateScheduleModal } from './private-schedule-modal'
+export { default as ScheduleBadge } from './schedule-badge'
+export { default as AppointmentBadge } from './appointment-badge'
