+++ src/entities/calendar/ui/private-schedule-badge.tsx
