+++ src/features/calendar/ui/index.ts
@@ -1,13 +1,6 @@
-export { default as CalendarCell } from './calendar-cell'
 export { default as Calendar } from './calendar'
-export { default as HeaderButton } from './header-button'
 export { default as HeaderTodayButton } from './header-today-button'
 export { default as HeaderLayout } from './header-layout'
-export { default as HeaderMonthLabel } from './header-month-label'
 export { default as InfiniteCalendar } from './infinite-calendar'
 export { default as MonthlyCalendar } from './monthly-calendar'
-export { default as PrivateScheduleBadge } from './private-schedule-badge'
-export { default as PrivateScheduleModal } from './private-schedule-modal'
-export { default as ScheduleBadge } from './schedule-badge'
-export { default as AppointmentBadge } from './appointment-badge'
 export { default as RenderScheduleBadges } from './render-schedule-badges'
